+++ lightswitch-capabilities/CHANGELOG.md
@@ -1,3 +1,7 @@
+Unreleased
+----------
+- Read pid namespace "level" of the userspace loader
+
 v0.3.0
 ------
 - Add feature detection for the `bpf_get_current_task_btf()` helper